ingestion/requirements.ingest.txt

pandas
requests
python-dotenv
azure-storage-blob
pyarrow